# docs/swagger.yaml
basePath: /v1
definitions:
...
  main.assignEmployeeRequest:
    properties:
      employee_id:
        type: integer
    type: object
  main.createScheduledShiftRequest:
    properties:
      employee_id:
        type: integer
      end_time:
        type: string
      notes:
        type: string
      role_id:
        type: integer
      shift_date:
        type: string
      shift_template_id:
        type: integer
      start_time:
        type: string
    type: object
  main.updateScheduledShiftRequest:
    properties:
      employee_id:
        type: integer
      end_time:
        type: string
      notes:
        type: string
      role_id:
        type: integer
      shift_date:
        type: string
      shift_template_id:
        type: integer
      start_time:
        type: string
    type: object
...
  store.ScheduledShift:
    properties:
      created_at:
        type: string
      employee_id:
        type: integer
      employee_name:
        description: Denormalized fields (stored in DB, synced via triggers)
        type: string
      end_time:
        type: string
      id:
        type: integer
      notes:
        type: string
      restaurant_id:
        type: integer
      role_color:
        type: string
      role_id:
        type: integer
      role_name:
        type: string
      schedule_id:
        type: integer
      shift_date:
        type: string
      shift_template_id:
        type: integer
      start_time:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for Restaurant Employee Scheduling Application (RESA)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: RESA API
paths:
...
  /restaurants/{restaurantID}/schedules/{scheduleID}/auto-populate:
    post:
      consumes:
      - application/json
      description: Creates scheduled shifts for all shift templates that don't have
        shifts yet
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurantID
        required: true
        type: integer
      - description: Schedule ID
        in: path
        name: scheduleID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Auto-populate schedule with template-based shifts
      tags:
      - scheduled-shifts
  /restaurants/{restaurantID}/schedules/{scheduleID}/shifts:
    get:
      consumes:
      - application/json
      description: Gets all scheduled shifts for a specific schedule
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurantID
        required: true
        type: integer
      - description: Schedule ID
        in: path
        name: scheduleID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.ScheduledShift'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: List all shifts for a schedule
      tags:
      - scheduled-shifts
    post:
      consumes:
      - application/json
      description: Creates a new scheduled shift for a specific schedule
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurantID
        required: true
        type: integer
      - description: Schedule ID
        in: path
        name: scheduleID
        required: true
        type: integer
      - description: Shift information
        in: body
        name: shift
        required: true
        schema:
          $ref: '#/definitions/main.createScheduledShiftRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.ScheduledShift'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create a new shift
      tags:
      - scheduled-shifts
  /restaurants/{restaurantID}/schedules/{scheduleID}/shifts/{shiftID}:
    delete:
      consumes:
      - application/json
      description: Deletes a scheduled shift by ID
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurantID
        required: true
        type: integer
      - description: Schedule ID
        in: path
        name: scheduleID
        required: true
        type: integer
      - description: Shift ID
        in: path
        name: shiftID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete a shift
      tags:
      - scheduled-shifts
    get:
      consumes:
      - application/json
      description: Gets a specific scheduled shift by ID
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurantID
        required: true
        type: integer
      - description: Schedule ID
        in: path
        name: scheduleID
        required: true
        type: integer
      - description: Shift ID
        in: path
        name: shiftID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.ScheduledShift'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get a shift
      tags:
      - scheduled-shifts
    patch:
      consumes:
      - application/json
      description: Updates an existing scheduled shift by ID
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurantID
        required: true
        type: integer
      - description: Schedule ID
        in: path
        name: scheduleID
        required: true
        type: integer
      - description: Shift ID
        in: path
        name: shiftID
        required: true
        type: integer
      - description: Updated shift information
        in: body
        name: shift
        required: true
        schema:
          $ref: '#/definitions/main.updateScheduledShiftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.ScheduledShift'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update a shift
      tags:
      - scheduled-shifts
  /restaurants/{restaurantID}/schedules/{scheduleID}/shifts/{shiftID}/assign:
    delete:
      consumes:
      - application/json
      description: Removes an employee assignment from a scheduled shift
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurantID
        required: true
        type: integer
      - description: Schedule ID
        in: path
        name: scheduleID
        required: true
        type: integer
      - description: Shift ID
        in: path
        name: shiftID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.ScheduledShift'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Unassign employee from shift
      tags:
      - scheduled-shifts
    patch:
      consumes:
      - application/json
      description: Assigns an employee to a scheduled shift
      parameters:
      - description: Restaurant ID
        in: path
        name: restaurantID
        required: true
        type: integer
      - description: Schedule ID
        in: path
        name: scheduleID
        required: true
        type: integer
      - description: Shift ID
        in: path
        name: shiftID
        required: true
        type: integer
      - description: Employee assignment information
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/main.assignEmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.ScheduledShift'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Assign employee to shift
      tags:
      - scheduled-shifts
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"